+++ test/variables.tf
@@ -1,4 +1,4 @@
 variable "cluster_vultr_api_key" {
-  type = string
+  type      = string
   sensitive = true
 }