+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Change Log
+## [v1.2.0](https://github.com/3letteragency/terraform-vultr-k0s/releases/tag/v1.2.0) (2021-06-12)
+### Features
+* Support K0s Helm deployments.
+### Changes
+* Convert module internal K0sctl configuration to HCL from YAML.
+### Fixes
+* Change Controller/Worker network interfaces from ens3/ens7 to enp1s0/enp6s0 due to Vultr image changes. 
+
 ## [v1.1.0](https://github.com/3letteragency/terraform-vultr-k0s/releases/tag/v1.1.0) (2021-06-06)
 ### Features
 * Write Kubeconfig locally option.